created mapTilesUtils.tsx
