+++ src/game/logic/boardLogic.ts
@@ -1,4 +1,4 @@
-import { GRID_SIZE } from "../conponents/board/Board";
+import { GRID_SIZE } from "../../utilities/globals";
 import { getCol, getRow, mapMatrix, rowFlip } from "./matrixUtils";
 
 export type Direction = "left" | "right" | "up" | "down";
