+++ src/utilities/globals.ts
@@ -1,0 +1,3 @@
+export const SWIPE_TIME = 1000;
+export const GRID_SIZE = 4;
+export const MARGIN_BETWEEN_TILES = 7;
